backend/requirements.txt
confluent-kafka>=2.3.0
fastapi>=0.109.2
redis>=5.0.1
pinecone-client>=3.1.0
openai>=1.12.0
anthropic>=0.18.1
supabase>=2.3.4
upstox-python-sdk>=2.0.0
yfinance>=0.2.36
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pdfplumber>=0.10.4
APScheduler>=3.10.4
SQLAlchemy>=2.0.25
asyncpg>=0.29.0
python-dotenv>=1.0.1
httpx>=0.25.2,<0.26.0
pandas>=2.2.0
numpy>=1.26.4
uvicorn>=0.27.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
celery>=5.3.6
